back_end/requirements.txt
pyppeteer
connexion==2.9.0
swagger-ui-bundle==0.0.9
flask==1.1.4
gunicorn==20.1.0
jsonpickle==2.0.0
markupsafe==2.0.1
flask-cors
Pillow
numpy